requirements.txt
telebot~=0.0.5
lyricsgenius~=3.0.1
python-dotenv~=1.0.0